[refactor]: Group 로직을 Room으로 이동 (#83)
