+++ app/src/main/java/com/texthip/thip/data/di/ServiceModule.kt
@@ -1,7 +1,6 @@
 package com.texthip.thip.data.di
 
 import com.texthip.thip.data.service.BookService
-import com.texthip.thip.data.service.GroupService
 import com.texthip.thip.data.service.RoomsService
 import com.texthip.thip.data.service.UserService
 import dagger.Module
@@ -14,12 +13,6 @@ import javax.inject.Singleton
 @Module
 @InstallIn(SingletonComponent::class)
 object ServiceModule {
-    
-    @Provides
-    @Singleton
-    fun provideGroupService(retrofit: Retrofit): GroupService {
-        return retrofit.create(GroupService::class.java)
-    }
 
     @Provides
     @Singleton
