+++ app/src/main/java/com/texthip/thip/data/model/rooms/request/RoomJoinRequest.kt
@@ -1,4 +1,4 @@
-package com.texthip.thip.data.model.group.request
+package com.texthip.thip.data.model.rooms.request
 
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
