+++ app/src/main/java/com/texthip/thip/data/repository/RoomsRepository.kt
@@ -1,11 +1,20 @@
 package com.texthip.thip.data.repository
 
+import com.texthip.thip.data.manager.GenreManager
+import com.texthip.thip.data.manager.UserDataManager
+import com.texthip.thip.data.manager.Genre
 import com.texthip.thip.data.model.base.handleBaseResponse
+import com.texthip.thip.data.model.rooms.request.CreateRoomRequest
+import com.texthip.thip.data.model.rooms.request.RoomJoinRequest
 import com.texthip.thip.data.model.rooms.request.RoomsCreateVoteRequest
 import com.texthip.thip.data.model.rooms.request.RoomsPostsLikesRequest
 import com.texthip.thip.data.model.rooms.request.RoomsRecordRequest
 import com.texthip.thip.data.model.rooms.request.RoomsVoteRequest
 import com.texthip.thip.data.model.rooms.request.VoteItem
+import com.texthip.thip.data.model.rooms.response.JoinedRoomListResponse
+import com.texthip.thip.data.model.rooms.response.MyRoomListResponse
+import com.texthip.thip.data.model.rooms.response.RoomMainList
+import com.texthip.thip.data.model.rooms.response.RoomRecruitingResponse
 import com.texthip.thip.data.service.RoomsService
 import javax.inject.Inject
 import javax.inject.Singleton
@@ -13,7 +22,77 @@ import javax.inject.Singleton
 @Singleton
 class RoomsRepository @Inject constructor(
     private val roomsService: RoomsService,
+    private val genreManager: GenreManager,
+    private val userDataManager: UserDataManager
 ) {
+    
+    /** 장르 목록 조회 */
+    fun getGenres(): Result<List<Genre>> {
+        return Result.success(genreManager.getGenres())
+    }
+    
+    /** 사용자 이름 조회(캐싱 데이터 사용)*/
+    fun getUserName(): Result<String> {
+        return Result.success(userDataManager.getUserName())
+    }
+    
+    /** 내가 참여 중인 모임방 목록 조회 */
+    suspend fun getMyJoinedRooms(page: Int): Result<JoinedRoomListResponse?> = runCatching {
+        val response = roomsService.getJoinedRooms(page)
+            .handleBaseResponse()
+            .getOrThrow()
+        
+        response?.let { joinedRoomsDto ->
+            userDataManager.cacheUserName(joinedRoomsDto.nickname)
+        }
+        
+        response
+    }
+
+    /** 카테고리별 모임방 섹션 조회 (마감임박/인기) */
+    suspend fun getRoomSections(genre: Genre? = null): Result<RoomMainList?> = runCatching {
+        val selectedGenre = genre ?: genreManager.getDefaultGenre()
+        val apiCategory = genreManager.mapGenreToApiCategory(selectedGenre)
+        
+        roomsService.getRooms(apiCategory)
+            .handleBaseResponse()
+            .getOrThrow()
+    }
+
+    /** 타입별 내 모임방 목록 조회 */
+    suspend fun getMyRoomsByType(type: String?, cursor: String? = null): Result<MyRoomListResponse?> = runCatching {
+        roomsService.getMyRooms(type, cursor)
+            .handleBaseResponse()
+            .getOrThrow()
+    }
+    
+    /** 모집중인 모임방 상세 정보 조회 */
+    suspend fun getRoomRecruiting(roomId: Int): Result<RoomRecruitingResponse> = runCatching {
+        roomsService.getRoomRecruiting(roomId)
+            .handleBaseResponse()
+            .getOrThrow()
+            ?: throw NoSuchElementException("모집중인 모임방 정보를 찾을 수 없습니다.")
+    }
+
+    /** 새 모임방 생성 */
+    suspend fun createRoom(request: CreateRoomRequest): Result<Int> = runCatching {
+        val response = roomsService.createRoom(request)
+            .handleBaseResponse()
+            .getOrThrow()
+            ?: throw NoSuchElementException("모임방 생성 응답을 받을 수 없습니다.")
+        response.roomId
+    }
+
+    /** 모임방 참여 또는 취소 */
+    suspend fun joinOrCancelRoom(roomId: Int, type: String): Result<String> = runCatching {
+        val request = RoomJoinRequest(type = type)
+        val response = roomsService.joinOrCancelRoom(roomId, request)
+            .handleBaseResponse()
+            .getOrThrow()
+            ?: throw NoSuchElementException("모임방 참여/취소 응답을 받을 수 없습니다.")
+        response.type
+    }
+
     suspend fun getRoomsPlaying(
         roomId: Int
     ) = runCatching {
