+++ app/src/main/java/com/texthip/thip/data/service/RoomsService.kt
@@ -1,10 +1,18 @@
 package com.texthip.thip.data.service
 
 import com.texthip.thip.data.model.base.BaseResponse
+import com.texthip.thip.data.model.rooms.request.CreateRoomRequest
+import com.texthip.thip.data.model.rooms.request.RoomJoinRequest
 import com.texthip.thip.data.model.rooms.request.RoomsCreateVoteRequest
 import com.texthip.thip.data.model.rooms.request.RoomsPostsLikesRequest
 import com.texthip.thip.data.model.rooms.request.RoomsRecordRequest
 import com.texthip.thip.data.model.rooms.request.RoomsVoteRequest
+import com.texthip.thip.data.model.rooms.response.CreateRoomResponse
+import com.texthip.thip.data.model.rooms.response.JoinedRoomListResponse
+import com.texthip.thip.data.model.rooms.response.MyRoomListResponse
+import com.texthip.thip.data.model.rooms.response.RoomJoinResponse
+import com.texthip.thip.data.model.rooms.response.RoomRecruitingResponse
+import com.texthip.thip.data.model.rooms.response.RoomMainList
 import com.texthip.thip.data.model.rooms.response.RoomsBookPageResponse
 import com.texthip.thip.data.model.rooms.response.RoomsCreateVoteResponse
 import com.texthip.thip.data.model.rooms.response.RoomsDeleteRecordResponse
@@ -22,6 +30,43 @@ import retrofit2.http.Path
 import retrofit2.http.Query
 
 interface RoomsService {
+
+    /** 참여 중인 모임방 목록 조회 */
+    @GET("rooms/home/joined")
+    suspend fun getJoinedRooms(
+        @Query("page") page: Int = 1
+    ): BaseResponse<JoinedRoomListResponse>
+
+    /** 카테고리별 모임방 목록 조회 (마감임박/인기) */
+    @GET("rooms")
+    suspend fun getRooms(
+        @Query("category") category: String = "문학"
+    ): BaseResponse<RoomMainList>
+
+    /** 내가 만든/참여한 모임방 목록 조회 */
+    @GET("rooms/my")
+    suspend fun getMyRooms(
+        @Query("type") type: String? = null,
+        @Query("cursor") cursor: String? = null
+    ): BaseResponse<MyRoomListResponse>
+
+    /** 모집중인 모임방 상세 정보 조회 */
+    @GET("rooms/{roomId}/recruiting")
+    suspend fun getRoomRecruiting(@Path("roomId") roomId: Int): BaseResponse<RoomRecruitingResponse>
+
+    /** 새 모임방 생성 */
+    @POST("rooms")
+    suspend fun createRoom(
+        @Body request: CreateRoomRequest
+    ): BaseResponse<CreateRoomResponse>
+
+    /** 모임방 참여/취소 */
+    @POST("rooms/{roomId}/join")
+    suspend fun joinOrCancelRoom(
+        @Path("roomId") roomId: Int,
+        @Body request: RoomJoinRequest
+    ): BaseResponse<RoomJoinResponse>
+
     @GET("rooms/{roomId}/playing")
     suspend fun getRoomsPlaying(
         @Path("roomId") roomId: Int
