+++ app/src/main/java/com/texthip/thip/ui/group/done/viewmodel/GroupDoneViewModel.kt
@@ -2,7 +2,7 @@ package com.texthip.thip.ui.group.done.viewmodel
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.texthip.thip.data.repository.GroupRepository
+import com.texthip.thip.data.repository.RoomsRepository
 import com.texthip.thip.ui.group.myroom.mock.RoomType
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.flow.MutableStateFlow
@@ -13,7 +13,7 @@ import javax.inject.Inject
 
 @HiltViewModel
 class GroupDoneViewModel @Inject constructor(
-    private val repository: GroupRepository
+    private val repository: RoomsRepository
 ) : ViewModel() {
 
     private val _uiState = MutableStateFlow(GroupDoneUiState())
