+++ app/src/main/java/com/texthip/thip/ui/group/myroom/component/GroupMainCard.kt
@@ -32,7 +32,7 @@ import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import coil.compose.AsyncImage
 import com.texthip.thip.R
-import com.texthip.thip.data.model.group.response.JoinedRoomResponse
+import com.texthip.thip.data.model.rooms.response.JoinedRoomResponse
 import com.texthip.thip.ui.theme.ThipTheme
 import com.texthip.thip.ui.theme.ThipTheme.colors
 import com.texthip.thip.ui.theme.ThipTheme.typography
